format code

+++ src/main/java/com/kmwllc/lucille/connector/xml/RecordingInputStream.java
@@ -1,6 +1,9 @@
 package com.kmwllc.lucille.connector.xml;
 
-import java.io.*;
+import java.io.ByteArrayOutputStream;
+import java.io.FilterInputStream;
+import java.io.IOException;
+import java.io.InputStream;
 
 // extended from example at:
 // https://bytes.com/topic/net/answers/818268-java-sax-parser-how-get-raw-xml-code-currently-parsingevent
@@ -27,8 +30,7 @@ public class RecordingInputStream extends FilterInputStream {
   @Override
   public synchronized int read(byte[] buf, int off, int len) throws IOException {
     int l = in.read(buf, off, len);
-    if (l == -1)
-      return -1;
+    if (l == -1) return -1;
     sink.write(buf, off, l);
     return l;
   }
@@ -44,9 +46,8 @@ public class RecordingInputStream extends FilterInputStream {
     int i = 0;
     byte[] buf = new byte[1024];
     while (l < len) {
-      i = read(buf, 0, (int) Math.min((long) buf.length, len - l));
-      if (i == -1)
-        break;
+      i = read(buf, 0, (int) Math.min(buf.length, len - l));
+      if (i == -1) break;
       l += i;
     }
     return l;
